toggle key modifers for keycodes combined/decorated with key modifer.
Right now keycodes combined with key modifiers produce the same keys regardless of the current state of the same key modifer.
This change adjust this behavior in such a way that active modifier keys inverse the modifiers of combined keycodes.
The idea behind this is to achive some kind of 'developer' layout, where all number keys are shifted via key modifer
and require the same modifer to be pressed to get an un-shifted key registered.

In my setup I put S(KB_1) ... S(KB_0) on the dedicated keymap layer and I want to get number key be entered via
pressing SHIFT + 1.